Introduce basic error handling infrastructure
This change introduces the basic error handling infrastructure to the
program. For lack of stabilization of the process::Termination we work
fairly hard to make do with just returning a Result from main. We
introduce an Error enum that can take represent every possible error
type the program can encounter. In addition to that, we provide the
WithCtx trait that can be used to add contextual information to an
error.