Add support for parsing public SSH keys
One of the first tasks the agent needs to accomplish is to respond to
requests detailing all the identities it knows about. Identity in this
sense is pretty much just a different term for a public key, which
uniquely identifies the identity a user has in a certain context. This
change marks a first step towards this goal. It adds support for parsing
public SSH keys. We use the ssh-keys crate to perform the actual task of
parsing the key into a suitable in-memory representation.