Add infrastructure for loading public keys from files
As a next step towards being able to serve requests asking for the
identities the agent knows about, this change adds support for reading
in public SSH key files into memory. This loading is done at startup of
the agent and right now we do not support late addition of public keys.
Note that while we identify public keys based on their extension
(*.pub), we do not unconditionally try to load all matching files.
Rather, we only load those for which a corresponding file with the .gpg
extension exists, because this agent only serves requests for PGP
encrypted private keys.