Add infrastructure for loading and decrypting private keys
Now that we can serve the known identities, the next step is for us to
actually to able to respond to signing requests. In order to accomplish
that we need to have the private keys corresponding to the public key in
question loaded.
This change adds the necessary infrastructure to load those keys. The
point of this agent is to be able to support PGP encrypted keys and we
use the gpgme crate to actually instruct PGP to decrypt the keys we
found.