Add support for serving signing requests
This change finally adds support for serving signing requests to the
agent. Responding to such a request with a proper signature of the
supplied payload is proof of possession of the private key and, hence,
the mechanism that grants access to a remote host.
In order to sign data using a private key we need yet another crate. We
decided to use ring, which aims to be easy to use and hard to misuse. It
is based in large parts on BoringSSL, a library that has proven itself
by now.