Add support for reading keys from a custom directory
So far we have unconditionally assumed SSH keys to be present in ~/.ssh.
Some users may, for whatever reason, prefer to store their keys in a
different directory.
To be more flexible in this respect, this change adds support for
parsing an optional argument to the program that represents the
directory the user wants the agent to read keys from.