Disable default features for env_logger
The agent's binary size is pretty significant at 1674720 bytes. In an
attempt to improve on the situation various crates were investigated. It
turns out that the env_logger dependency pulls in the regex crate which
in turn causes a huge binary size increase.
This change disables the default features for the env_logger crate,
which removes the dependency to regex. Doing so gets the binary size
down to 880064 bytes.