Do not cache public keys
This change makes the agent completely stateless: instead of keeping a
list of public keys and working on this list, we now load public keys on
demand as well. As loading those keys does not involve any decryption,
the operation is typically very fast. The upside of doing so is that we
are now able to detect newly added keys. That is, while in the past the
agent had to be restarted for newly added keys to be recognized, those
are now trivially picked up.
In order to implement detection of newly added key material in the old
scheme we would have to opt for an inotify-like mechanism, which just
adds implementation complexity for little benefit.