Refactor signing code in preparation for RSA support
The PrivateKey::sign method currently uses a match statement that
evaluates to an ed25519 key pair. Once we add support for RSA in a
different arm, returning such a specific key pair will no longer work.
Hence, with this change we refactor the code in such a way that we have
dedicated signature functions for each key type which only return a
signature blob.